website/requirements.txt
numpy==1.23.4
pandas==1.5.1
opencv-python==4.6.0.66
Pillow==9.2.0
requests==2.28.1